BipEp/requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.4
django-crispy-forms==1.10.0
django-environ==0.4.5
django-filter==2.4.0
django-mysql==3.10.0
django-session-timeout==0.1.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
PyJWT==1.7.1
six==1.15.0
